Finalize directory structure Remove unrelated classes from Card.{cpp,hpp}
