+++ include/kabufuda/WideStringConvert.hpp
@@ -1,3 +1,6 @@
+#ifndef __KABU_WIDESTRINGCONVERT_HPP__
+#define __KABU_WIDESTRINGCONVERT_HPP__
+
 #include <string>
 
 namespace kabufuda
@@ -5,3 +8,5 @@ namespace kabufuda
 std::string WideToUTF8(const std::wstring& src);
 std::wstring UTF8ToWide(const std::string& src);
 }
+
+#endif // __KABU_WIDESTRINGCONVERT_HPP__
