+++ lib/kabufuda/Card.cpp
@@ -493,241 +493,4 @@ Card::operator bool() const
 
     return true;
 }
-
-File::File(char data[])
-{
-    memcpy(__raw, data, 0x40);
-}
-
-File::File(const char* filename)
-{
-    memset(__raw, 0xFF, 0x40);
-    memset(m_filename, 0, 32);
-    size_t len = strlen(filename);
-    len = std::min<size_t>(len, 32);
-    memcpy(m_filename, filename, len);
-}
-void File::swapEndian()
-{
-    m_modifiedTime = SBig(m_modifiedTime);
-    m_imageOffset = SBig(m_imageOffset);
-    m_iconFmt = SBig(m_iconFmt);
-    m_animSpeed = SBig(m_animSpeed);
-    m_firstBlock = SBig(m_firstBlock);
-    m_blockCount = SBig(m_blockCount);
-    m_reserved2 = SBig(m_reserved2);
-    m_commentAddr = SBig(m_commentAddr);
-}
-
-File::File()
-{
-    memset(__raw, 0xFF, 0x40);
-}
-
-void BlockAllocationTable::swapEndian()
-{
-    m_checksum = SBig(m_checksum);
-    m_checksumInv = SBig(m_checksumInv);
-    m_updateCounter = SBig(m_updateCounter);
-    m_freeBlocks = SBig(m_freeBlocks);
-    m_lastAllocated = SBig(m_lastAllocated);
-    std::for_each(std::begin(m_map), std::end(m_map), [](uint16_t& val){
-        val = SBig(val);
-    });
-}
-
-void BlockAllocationTable::updateChecksum()
-{
-    swapEndian();
-    calculateChecksumBE(reinterpret_cast<uint16_t*>(__raw + 4), 0xFFE, &m_checksum, &m_checksumInv);
-    swapEndian();
-}
-
-bool BlockAllocationTable::valid() const
-{
-    uint16_t ckSum, ckSumInv;
-    BlockAllocationTable tmp = *this;
-    tmp.swapEndian();
-    calculateChecksumBE(reinterpret_cast<const uint16_t*>(tmp.__raw + 4), 0xFFE, &ckSum, &ckSumInv);
-    return (SBig(ckSum) == m_checksum && SBig(ckSumInv) == m_checksumInv);
-}
-
-BlockAllocationTable::BlockAllocationTable(uint32_t blockCount)
-{
-    memset(__raw, 0, BlockSize);
-    m_freeBlocks = uint16_t(blockCount - FSTBlocks);
-    m_lastAllocated = 4;
-    updateChecksum();
-}
-
-uint16_t BlockAllocationTable::getNextBlock(uint16_t block) const
-{
-    if ((block < FSTBlocks) || (block > (BATSize - FSTBlocks)))
-        return 0;
-
-    return m_map[block - FSTBlocks];
-}
-
-uint16_t BlockAllocationTable::nextFreeBlock(uint16_t maxBlock, uint16_t startingBlock) const
-{
-    if (m_freeBlocks > 0)
-    {
-        maxBlock = std::min(maxBlock, uint16_t(BATSize));
-        for (uint16_t i = startingBlock; i < maxBlock; ++i)
-            if (m_map[i - FSTBlocks] == 0)
-                return i;
-
-        for (uint16_t i = FSTBlocks; i < startingBlock; ++i)
-            if (m_map[i - FSTBlocks] == 0)
-                return i;
-    }
-
-    return 0xFFFF;
-}
-
-bool BlockAllocationTable::clear(uint16_t first, uint16_t count)
-{
-    std::vector<uint16_t> blocks;
-    while (first != 0xFFFF && first != 0)
-    {
-        blocks.push_back(first);
-        first = getNextBlock(first);
-    }
-    if (first > 0)
-    {
-        size_t length = blocks.size();
-        if (length != count)
-            return false;
-
-        for (size_t i= 0 ; i < length ; ++i)
-            m_map[blocks.at(i) - FSTBlocks] = 0;
-        m_freeBlocks += count;
-        return true;
-    }
-    return false;
-}
-
-uint16_t BlockAllocationTable::allocateBlocks(uint16_t count, uint16_t maxBlocks)
-{
-    uint16_t firstBlock = nextFreeBlock(maxBlocks - FSTBlocks, m_lastAllocated + 1);
-    uint16_t freeBlock = firstBlock;
-    if (freeBlock != 0xFFFF)
-    {
-        uint16_t tmpCount = count;
-        while ((count--) > 0)
-        {
-            m_map[(freeBlock - FSTBlocks)] = 0xFFFF;
-            if (count != 0)
-            {
-                m_map[(freeBlock - FSTBlocks)] = nextFreeBlock(maxBlocks - FSTBlocks, m_lastAllocated + 1);
-                freeBlock = m_map[(freeBlock - FSTBlocks)];
-            }
-            m_lastAllocated = freeBlock;
-        }
-
-        m_freeBlocks -= tmpCount;
-    }
-    return firstBlock;
-}
-
-void Directory::swapEndian()
-{
-    std::for_each(std::begin(m_files), std::end(m_files), [](File& f){
-        f.swapEndian();
-    });
-
-    m_updateCounter = SBig(m_updateCounter);
-    m_checksum = SBig(m_checksum);
-    m_checksumInv = SBig(m_checksumInv);
-}
-
-void Directory::updateChecksum()
-{
-    swapEndian();
-    calculateChecksumBE(reinterpret_cast<uint16_t*>(__raw), 0xFFE, &m_checksum, &m_checksumInv);
-    swapEndian();
-}
-
-bool Directory::valid() const
-{
-    uint16_t ckSum, ckSumInv;
-    Directory tmp = *this;
-    tmp.swapEndian();
-    calculateChecksumBE(reinterpret_cast<const uint16_t*>(tmp.__raw), 0xFFE, &ckSum, &ckSumInv);
-    return (SBig(ckSum) == m_checksum && SBig(ckSumInv) == m_checksumInv);
-}
-
-Directory::Directory()
-{
-    memset(__raw, 0xFF, BlockSize);
-    m_updateCounter = 0;
-    updateChecksum();
-}
-
-Directory::Directory(uint8_t data[])
-{
-    memcpy(__raw, data, BlockSize);
-}
-
-Directory::Directory(const Directory& other)
-{
-    memcpy(__raw, other.__raw, BlockSize);
-}
-
-void Directory::operator=(const Directory& other)
-{
-    memcpy(__raw, other.__raw, BlockSize);
-}
-
-File* Directory::getFirstFreeFile(const char* game, const char* maker, const char* filename)
-{
-    for (uint16_t i = 0 ; i < 127 ; i++)
-    {
-        if (m_files[i].m_id[0] == 0xFF)
-        {
-            File* ret = &m_files[i];
-            *ret = File(filename);
-            if (game && strlen(game) == 4)
-                memcpy(ret->m_id, game, 4);
-            if (maker && strlen(maker) == 2)
-                memcpy(ret->m_maker, maker, 2);
-            return ret;
-        }
-    }
-
-    return nullptr;
-}
-
-File* Directory::getFile(const char* game, const char* maker, const char* filename)
-{
-    for (uint16_t i = 0 ; i < 127 ; i++)
-    {
-        if (game && strlen(game) == 4 && memcmp(m_files[i].m_id, game, 4))
-            continue;
-        if (maker && strlen(maker) == 2 && memcmp(m_files[i].m_maker, maker, 2))
-            continue;
-        if (!strcmp(m_files[i].m_filename, filename))
-            return &m_files[i];
-    }
-
-    return nullptr;
-}
-
-void calculateChecksumBE(const uint16_t* data, size_t len, uint16_t* checksum, uint16_t* checksumInv)
-{
-    *checksum = 0;
-    *checksumInv = 0;
-    for (size_t i = 0; i < len; ++i)
-    {
-        *checksum += SBig(data[i]);
-        *checksumInv += SBig(uint16_t(data[i] ^ 0xFFFF));
-    }
-
-    *checksum = SBig(*checksum);
-    *checksumInv = SBig(*checksumInv);
-    if (*checksum == 0xFFFF)
-        *checksum = 0;
-    if (*checksumInv == 0xFFFF)
-        *checksumInv = 0;
-}
 }
