+++ lib/kabufuda/Util.cpp
@@ -23,4 +23,22 @@ uint64_t getGCTime()
 
     return (uint64_t)(sysTime + tzDiff + tzDST) - 0x386D4380;
 }
+
+void calculateChecksumBE(const uint16_t* data, size_t len, uint16_t* checksum, uint16_t* checksumInv)
+{
+    *checksum = 0;
+    *checksumInv = 0;
+    for (size_t i = 0; i < len; ++i)
+    {
+        *checksum += SBig(data[i]);
+        *checksumInv += SBig(uint16_t(data[i] ^ 0xFFFF));
+    }
+
+    *checksum = SBig(*checksum);
+    *checksumInv = SBig(*checksumInv);
+    if (*checksum == 0xFFFF)
+        *checksum = 0;
+    if (*checksumInv == 0xFFFF)
+        *checksumInv = 0;
+}
 }
