+++ lib/kabufuda/WideStringConvert.cpp
